% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{annotate_methylation}
\alias{annotate_methylation}
\title{Annotate methylation data}
\usage{
annotate_methylation(
  x,
  annotation,
  mult = "first",
  annotation_col = NULL,
  keep_annotation = FALSE
)
}
\arguments{
\item{x}{data.table with methylation information.}

\item{annotation}{data.table with annotation information}

\item{mult}{how to handle multi overlaps. See \link[data.table:foverlaps]{data.table::foverlaps}
for details.}

\item{annotation_col}{optional, column name of annotation to add to x. If
null (the default) the entire table is added.}

\item{keep_annotation}{optional logical, should the start and end columns of
the annotation be retained? If TRUE, these columns are named feature_start and
feature_end, or, if annotation_col is specified, that name is used instead of
feature.}
}
\value{
data.table x annotated with the information in annotation
}
\description{
Annotate methylation data
}
\details{
x and annotation arguments must contain columns named seqnames,
start and end. If the elements in annotation overlap, the overlaps are handled
using the mult argument. Defaults to "first" which means only the first feature
in annotation is used. Alternatively "all" can be used, but that leads to a
CpG being tested multiple times.
}
