% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{apply_by_line}
\alias{apply_by_line}
\title{Apply function line by line}
\usage{
apply_by_line(file, fun, ..., chunks = 10000)
}
\arguments{
\item{file}{path to file to be processed}

\item{fun}{function to be applied to line}

\item{...}{other parameters applied to function}
}
\description{
Apply function line by line
}
