% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq_updated.R
\name{catHeader}
\alias{catHeader}
\title{Generate appropriate tab-header strings.
Useful in a loop to contain each plot in its own tab.}
\usage{
catHeader(text = "", level = 3)
}
\arguments{
\item{text}{The text in the header}

\item{level}{the level of the header, use one less than parent}
}
\description{
Generate appropriate tab-header strings.
Useful in a loop to contain each plot in its own tab.
}
