% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_bulk_RNA_seq.R
\name{catHeader_w_tabset}
\alias{catHeader_w_tabset}
\title{Generate appropriate tab-header strings with .tabset attribute
Useful in a loop to contain each plot in its own tab.}
\usage{
catHeader_w_tabset(text = "", level = 3)
}
\arguments{
\item{text}{The text in the header}

\item{level}{the level of the header, use one less than parent}
}
\description{
Generate appropriate tab-header strings with .tabset attribute
Useful in a loop to contain each plot in its own tab.
}
