% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{computer_friendly_names}
\alias{computer_friendly_names}
\title{Metadata column names conversion list}
\format{
An object of class \code{character} of length 27.
}
\usage{
computer_friendly_names
}
\description{
List of conversions of commonly used human readable metadata column names and
computer friendly alternatives
}
\keyword{datasets}
