% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cov_list}
\alias{cov_list}
\title{Small example of a list of coverage files}
\format{
An object of class \code{list} of length 6.
}
\usage{
cov_list
}
\description{
A dataset with a small toy example of what a list of Bismark cov files look
like
}
\keyword{datasets}
