% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{covfile_reader}
\alias{covfile_reader}
\title{Read cov file and optionally aggregate to known CpGs}
\usage{
covfile_reader(file, cpgs = NULL)
}
\arguments{
\item{file}{character path to cov file}

\item{cpgs}{data.table containing the position of known CpGs}
}
\value{
data.table with methylation information, optinally aggregated to known CpGs
}
\description{
Read cov file and optionally aggregate to known CpGs
}
