% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisulfite_analysis.R
\name{edger_for_rrbs}
\alias{edger_for_rrbs}
\title{Wrapper for running RRBS analysis through edgeR}
\usage{
edger_for_rrbs(y, design, parallel = TRUE)
}
\arguments{
\item{y}{DGEList with a column for unmethylated and a column for methylated counts per sample.}

\item{design}{design as generated by model.matrix. modelMatrixMeth is called in this function.}

\item{parallel}{logical, should large datasets be run in parallel. If true, samples with more than
100.000 rows are split into 50.000 row chunks and analyzed separately.}

\item{aggregate_by}{optional, name of column in y$genes to aggregate by.}

\item{sample_pattern}{}
}
\value{

}
\description{
Wrapper for running RRBS analysis through edgeR
}
