% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_analysis.R
\name{edgeR_tester}
\alias{edgeR_tester}
\title{Wrapper for testing coefficients or contrasts from an edgeR model}
\usage{
edgeR_tester(
  contrast = NULL,
  coef = NULL,
  efit,
  contrast_matrix,
  id_name = "ENSEMBL"
)
}
\arguments{
\item{contrast}{character, name of contrast to be tested}

\item{coef}{character, name of coefficient to be tested}

\item{efit}{DGEGLM object, object as created by edgeR function glmQLFit or glmFit}

\item{contrast_matrix}{matrix, typically created by the edgeR function makeContrasts}

\item{id_name}{character, name of the gene ids}
}
\value{
data.table with edgeR results
}
\description{
Wrapper for testing coefficients or contrasts from an edgeR model
}
\examples{
\dontrun{
y <- prepare_rna_dgelist(rna_counts, example_metadata, sample_col = "Sample.ID")
design <- model.matrix(~ 0 + Genotype + Batch, data = y$samples)
y <- edgeR::estimateDisp(y, design = design)
rownames(y) <- y$genes$Geneid
contrast_matrix <- limma::makeContrasts(Genotype = GenotypeKO - GenotypeWT, 
                                        levels = y$design)
efit <- glmQLFit(y, design = y$design, robust = TRUE)
edgeR_tester("Genotype", efit = efit, contrast_matrix = contrast_matrix)
}
}
