% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_analysis.R
\name{enrich_test_wrapper}
\alias{enrich_test_wrapper}
\title{Shallow wrapper to reformat fry/camera output}
\usage{
enrich_test_wrapper(contrast, index, y, fun, ...)
}
\arguments{
\item{contrast}{name of contrast to be tested or numeric vector of same
length as the number of columns of design. See edgeR::camera.}

\item{index}{list of index vectors. See edgeR::camera.}

\item{y}{DGEList object.}

\item{fun}{function to use for testing. Typically limma::camera or limma::fry}

\item{...}{other parameters passed to fun.}
}
\value{
data.table with enrichment test results.
}
\description{
Shallow wrapper to reformat fry/camera output
}
