% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_metadata}
\alias{example_metadata}
\title{Small example of a metadata}
\format{
An object of class \code{data.frame} with 6 rows and 4 columns.
}
\usage{
example_metadata
}
\description{
A small example of a metadata sheet
}
\keyword{datasets}
