% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{exponent_format}
\alias{exponent_format}
\title{format numbers for exponent}
\usage{
exponent_format()
}
\value{
function that converts numbers to 10^x format
}
\description{
format numbers for exponent
}
\examples{
exponent_format()
}
