% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_analysis.R
\name{featurecounts_reader}
\alias{featurecounts_reader}
\title{Prepare featureCounts output files for RNA-seq analysis}
\usage{
featurecounts_reader(x, regex)
}
\arguments{
\item{regex}{optional regular expression. Default option works with our pipeline}

\item{files}{list of featureCounts output files}
}
\value{
data.table of gene counts and formatted gene information. FeatureCounts
reports chromosome, start, end and strand for each exon. These are reformatted
so start contains the start of the first exon and end the end of the last exon.
Only one copy of chromosome and strand are reported. Column names are also
reformatted to match with Bioconductor nomenclature (seqnames instead of Chr, etc.).
If a regex is supplied it is used to remove text from the sample names, if not
specified a regex that works with Ali's pipeline is used.
}
\description{
Prepare featureCounts output files for RNA-seq analysis
}
\examples{
\dontrun{
 featurecount_files <- list.files(path = "path/to", pattern = "files", full.names = TRUE)
 count_matrix <- merge_featurecounts(featurecount_files)
}
}
