% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{filter_methylation}
\alias{filter_methylation}
\title{Filter methylation data}
\usage{
filter_methylation(x, remove = NULL, select = NULL)
}
\arguments{
\item{x}{data.table with methylation information.}

\item{remove}{optional data.table with regions to be removed}

\item{select}{optional data.table with regions to be selected}
}
\value{
data.table filtered based on the tables in remove and select
}
\description{
Filter methylation data
}
\details{
all three arguments must contain columns named seqnames, start and end.
If regions are present in both remove and select, they are removed.
}
