% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{find_pvalue_cutoff}
\alias{find_pvalue_cutoff}
\title{Find P-value cutoff}
\usage{
find_pvalue_cutoff(x, cutoff = 0.05)
}
\arguments{
\item{x}{vector of p-values}

\item{cutoff}{desired FDR cutoff}
}
\description{
Find P-value cutoff
}
\details{
This function takes a list of P-values, and calculates what the first P-value
above a certain FDR threshold would have had to be in order to have an FDR
less than said threshold. Uses an iterative approach to approximate, but can
be made arbitrarily precise.
Designed to be used when generating volcano-plots
}
