% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fix_nonbreaking_space}
\alias{fix_nonbreaking_space}
\title{Remove non-breaking space}
\usage{
fix_nonbreaking_space(x)
}
\arguments{
\item{x}{string}
}
\value{
string, where non-breaking space has been replaced with regular space
}
\description{
Remove non-breaking space
}
