% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{genomic_log_trans}
\alias{genomic_log_trans}
\title{Genomic log transformation}
\usage{
genomic_log_trans()
}
\value{
transformer, see scales::trans_new
}
\description{
Genomic log transformation
}
\details{
Returns a transformer that nicely handles log10 scale in a genomic context.
}
