% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq.R
\name{get_DE_datatable}
\alias{get_DE_datatable}
\title{Return a nicely formatted DT::datatable object of the DE results}
\usage{
get_DE_datatable(df, type = "edgeR")
}
\arguments{
\item{df}{The dataframe with the DE results}

\item{type}{Whether the results is edgeR or voom}
}
\value{
A DT::datatable
}
\description{
Return a nicely formatted DT::datatable object of the DE results
}
\examples{
get_DE_datatable()
}
