% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_analysis.R
\name{get_enrichment_terms}
\alias{get_enrichment_terms}
\title{Get terms for gene ontology enrichment analysis.}
\usage{
get_enrichment_terms(
  org_db,
  ensembl_ids,
  min_genes = 5,
  max_genes = 500,
  cache_path
)
}
\arguments{
\item{org_db}{AnnotationDb class object matching the species analyzed}

\item{ensembl_ids}{optional character vector. ENSEMBL genes to subset}

\item{min_genes}{integer, minimum number of genes to include term}

\item{max_genes}{integer, maximum number of genes to include term}

\item{cache_path}{character, optional path to cached Reactome file}
}
\value{
list of data formatted to work with the camera and fry wrappers in this package.
}
\description{
Get terms for gene ontology enrichment analysis.
}
\examples{
\dontrun{
library(org.Mm.eg.db)
#rna_counts is an example dataset included in the package
get_enrichment_terms(org.Mm.eg.db, rna_counts$Geneid)
}
}
