% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq.R
\name{get_GO_datatable}
\alias{get_GO_datatable}
\title{Return a nicely formatted DT::datatable of the GO-results}
\usage{
get_GO_datatable(df)
}
\arguments{
\item{df}{The GO-results}
}
\value{
A DT::datatable
}
\description{
Return a nicely formatted DT::datatable of the GO-results
}
\examples{
get_GO_datatable()
}
