% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq_updated.R
\name{get_GO_datatable_updated}
\alias{get_GO_datatable_updated}
\title{*** CANGED: NOW PASSES FURTHER ARGUMENTS TO DATATABLE, AND
USES SET PIXEL HEIGHT TO AVOID PROBLEMS IN LOOPS ****
Return a nicely formatted DT::datatable of the GO-results}
\usage{
get_GO_datatable_updated(df, ...)
}
\arguments{
\item{df}{The GO-results}

\item{...}{extra arguments passed to DT::datatable()}
}
\value{
A DT::datatable
}
\description{
*** CANGED: NOW PASSES FURTHER ARGUMENTS TO DATATABLE, AND
USES SET PIXEL HEIGHT TO AVOID PROBLEMS IN LOOPS ****
Return a nicely formatted DT::datatable of the GO-results
}
\examples{
get_GO_datatable()
}
