% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_analysis.R
\name{get_go_terms}
\alias{get_go_terms}
\title{Get ENSEMBL to GO data}
\usage{
get_go_terms(org_db, gene_ids, gene_id_key_type)
}
\arguments{
\item{org_db}{AnnotationDb class object matching the species analyzed}

\item{ensembl_ids}{optional character vector. ENSEMBL genes to subset}
}
\value{
list of data.table with ENSEMBL to GO mappings
}
\description{
Get ENSEMBL to GO data
}
\details{
Prepare ENSEMBL to GO mappings. If ensembl_ids are specified only
terms relating to these genes are returned.
}
\examples{
\dontrun{
library(org.Mm.eg.db)
get_go_terms(org.Mm.eg.db)
}
}
