% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisulfite_analysis.R
\name{get_groupwise_methylation}
\alias{get_groupwise_methylation}
\title{Get groupwise average methylation metric}
\usage{
get_groupwise_methylation(methylation, design)
}
\arguments{
\item{methylation}{matrix or object coercible to a matrix containing methylation metrics}

\item{design}{design matrix with experimental design}
}
\value{
matrix with groupwise averages of methylation metrics
}
\description{
Get the average groupwise methylation metric (in case of a '~ 0 + group' desing)
or effect size (in case of a '~ a * b' model).
}
\examples{
set.seed(1)
n_samples <- 4
n_genes <- 10
counts <- matrix(sample(1:20, n_samples * 2 * n_genes, replace = TRUE), 
                 ncol = n_samples * 2)
colnames(counts) <- paste0("sample", rep(1:4, each = 2), c("_Me", "_Un"))
methylation_levels <- get_methylation_metric(counts)

group <- c("a", "a", "b", "b")
design <- model.matrix(~ 0 + group)

get_groupwise_methylation(methylation_levels, design)
}
