% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_figures.R
\name{get_MD_data_for_sample}
\alias{get_MD_data_for_sample}
\title{Get MD data for a single sample}
\usage{
get_MD_data_for_sample(
  object,
  column,
  xlab = "Average log CPM (this sample and others)",
  ylab = "log-ratio (this sample vs others)",
  prior.count = 3
)
}
\arguments{
\item{object}{The DGElist object with transcript/gene counts and sample
information encoded in the y$sample object}
}
\value{
A long tibble with two columns: mean and diff values
}
\description{
Get MD data for a single sample
}
