% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataFromLabGuru.R
\name{get_metadata}
\alias{get_metadata}
\title{Download metadata from labguru}
\usage{
get_metadata(exp_id = NULL, exp_name = NULL)
}
\arguments{
\item{exp_id}{numeric experimental id, fastest way to get data}

\item{exp_name}{character of experimental name}
}
\description{
Download metadata from labguru when given a experiment id or
an experiment name
}
\examples{
\dontrun{
labguru_set_token(
  token = "abcdefg",
  server = "https://sund.labguru.com"
)
get_metadata(exp_id = 1)
}

}
