% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisulfite_analysis.R
\name{get_methylation_metric}
\alias{get_methylation_metric}
\alias{get_methylation_metric.default}
\alias{get_methylation_metric.data.table}
\alias{get_methylation_metric.DGEList}
\title{Get methylation metric}
\usage{
get_methylation_metric(x, metric = "percent")

\method{get_methylation_metric}{default}(x, metric = "percent")

\method{get_methylation_metric}{data.table}(x, metric = "percent")

\method{get_methylation_metric}{DGEList}(x, metric = "percent")
}
\arguments{
\item{x}{object to calculate methylation metric from}

\item{metric}{string, method to use, currently only percent and Mvalue are supported}
}
\value{
object of the same type as x, except for DGEList which returns an object of
the same type as x$counts (typically a matrix).
}
\description{
Get various per sample methylation metrics. Currently only percent methylation
and Mvalue (methylated counts / unmethylated counts) are supported.
}
\details{
x can be a matrix, a data.frame, a data.table or an edgeR DGEList.

x must contain only columns with counts and each sample must have two columns,
one ending in '_Me' and one ending in '_Un' with methylated and unmethylated
counts respectively. The columns need not be in that order.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Methylation metric default

\item \code{data.table}: Methylation metric for data.tables

\item \code{DGEList}: Methylation metric for DGELists
}}

\examples{
set.seed(1)
n_samples <- 4
n_genes <- 10
counts <- matrix(sample(1:20, n_samples * 2 * n_genes, replace = TRUE), 
                 ncol = n_samples * 2)
colnames(counts) <- paste0("sample", rep(1:4, each = 2), c("_Me", "_Un"))
get_methylation_metric(counts)
get_methylation_metric(counts, metric = "Mvalue")
}
