% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_analysis.R
\name{get_reactome_terms}
\alias{get_reactome_terms}
\title{Get ENSEMBL to Reactome data}
\usage{
get_reactome_terms(org_name, gene_ids, cache_path = NULL, gene_id_key_type)
}
\arguments{
\item{org_name}{optional string, organism name to subset. Common options are "Mus musculus" and "Homo sapiens"}

\item{cache_path}{optional path to cache}

\item{ensembl_ids}{optional character vector. ENSEMBL genes to subset}
}
\value{
data.table with ENSEMBL to Reactome mapping
}
\description{
Get ENSEMBL to Reactome data
}
\details{
Load ENSEMBL to Reactome mappings. If org_name is specified only terms
with that species is returned. If ensembl_ids are specified only terms relating to
these genes are returned.

If cache_path is set, the specified file
is loaded if it exists. If it does not exist it is downloaded and the saved to that
location.
}
\examples{
\dontrun{
get_reactome_terms("Homo sapiens")
}
}
