% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq_updated.R
\name{get_sig_entities_between_tissues_across_conds}
\alias{get_sig_entities_between_tissues_across_conds}
\title{Gets a nested list of significant entities (genes or GO-terms) between tissues across contrasts.
Assumes there is a data dir with each tissue described in its name, with GO and DE-tables in them
with corresponding names.}
\usage{
get_sig_entities_between_tissues_across_conds(
  type,
  cond_str,
  tissue_str,
  genes_type
)
}
\arguments{
\item{type}{Either "DE" or "GO" depending on the type of entity}

\item{cond_str}{Character vector: Comma-separated list of contrasts as present in the data directory, i.e. Group2,Group3}

\item{tissue_str}{Character vector: Comma-seperated list of tissues as present in the project root, i.e. Hyp,NAc}
}
\value{
A nested list, where the upper levels corresponds to supplied tissues,
and the lower levels each contain a vector of significant gene names or GO-terms for the given condition
}
\description{
Gets a nested list of significant entities (genes or GO-terms) between tissues across contrasts.
Assumes there is a data dir with each tissue described in its name, with GO and DE-tables in them
with corresponding names.
}
