% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq_updated.R
\name{get_sig_entities_from_df}
\alias{get_sig_entities_from_df}
\title{Get number of significant DEGs or GO-terms from dataframe}
\usage{
get_sig_entities_from_df(df, genes_type = "NCBI", type = "DE")
}
\arguments{
\item{df}{Data frame as produced by get_GO/DE_datatable}

\item{genes_type}{character vector, NCBI or ENSEMBL}

\item{type}{DE or GO}
}
\value{
character vector: gene names or ensembl IDs with FDR<0.5
}
\description{
Get number of significant DEGs or GO-terms from dataframe
}
