% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_bulk_RNA_seq.R
\name{get_sig_entities_from_path}
\alias{get_sig_entities_from_path}
\title{Get number of significant DEGs or GO-terms from tsv file as produced by get_DE_datatable}
\usage{
get_sig_entities_from_path(path, type, genes_type)
}
\arguments{
\item{path}{character vector: Path to the .tsv file}

\item{type}{character vector: DE or GO}

\item{genes_type}{character vector: DE or GO}
}
\value{
character vector: gene names or ensembl IDs with FDR<0.5
}
\description{
Get number of significant DEGs or GO-terms from tsv file as produced by get_DE_datatable
}
