% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataFromLabGuru.R
\name{getMetadata}
\alias{getMetadata}
\title{Download metadata from labguru}
\usage{
getMetadata(expId = NULL, expName = NULL)
}
\arguments{
\item{expId}{numeric experimental id, fastest way to get data}

\item{expName}{character of experimental name}
}
\description{
Download metadata from labguru when given a experiment id or
an experiment name
}
\examples{
\dontrun{
labguru_set_token(token = "abcdefg",
server = "https://sund.labguru.com")
getMetadata(expId = 1)
}
}
