% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_figures.R
\name{ggplot_MD}
\alias{ggplot_MD}
\title{Plot MD-figures for all samples in DGElist}
\usage{
ggplot_MD(object)
}
\arguments{
\item{object}{The DGElist object with transcript/gene counts and sample information
encoded in the y$sample object}
}
\value{
A facetted ggplot
}
\description{
Plot MD-figures for all samples in DGElist
}
\examples{
ggplot_MD()
}
