% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_figures.R
\name{ggplot_mds_repel}
\alias{ggplot_mds_repel}
\title{Plot MDS for a given dimensions and color it}
\usage{
ggplot_mds_repel(y, dims, color_by)
}
\arguments{
\item{y}{The DGElist object with transcript/gene counts and sample information}

\item{dims}{A numeric vector of the two dimensions of the MDS plot}

\item{color_by}{Character vector of the colour to color the points and labels by,
encoded in the y$sample object}
}
\value{
The ggplot-object with title, labels, and appropriate color
}
\description{
Plot MDS for a given dimensions and color it
}
\examples{
ggplot_mds_repel()
}
