% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq_updated.R
\name{ggplot_mds_repel_updated}
\alias{ggplot_mds_repel_updated}
\title{Plot samples in two-dimensional space using MDS}
\usage{
ggplot_mds_repel_updated(y, dims, color_by)
}
\arguments{
\item{y}{The DGE list object}

\item{dims}{The dimensions to plot as a vector}

\item{color_by}{The vector to colour the plots by}
}
\value{
Returns a list of ggplots coloured according to color_by
}
\description{
Plot samples in two-dimensional space using MDS
}
