% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_bulk_RNA_seq.R
\name{ggplot_volcano_updated}
\alias{ggplot_volcano_updated}
\title{Make a volcano plot using ggplot}
\usage{
ggplot_volcano_updated(
  df,
  plot_genes = "all",
  highlight_genes = c("interesting_gene_name_1", "interesting_gene_name_2"),
  only_FDR_sig = FALSE,
  only_nom_sig = FALSE,
  interactive_plot = TRUE,
  title = ""
)
}
\arguments{
\item{df}{gene results from EdgeR or voom}

\item{interactive_plot}{Whether or not to print a plotly interactive plot}

\item{genes}{character vector, "all" for plotting all genes,
otherwise only plot vector of genes}

\item{only_sig}{Whether or not only to print FDR significant genes}
}
\value{
The ggplot object
}
\description{
Make a volcano plot using ggplot
}
