% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian_analysis.R
\name{limorhyde_harmonics}
\alias{limorhyde_harmonics}
\title{Limorhyde function with multiple harmonics}
\usage{
limorhyde_harmonics(
  df,
  time_colname,
  harmonics = 1,
  period = 24,
  sinusoid = TRUE,
  n_knots = 3
)
}
\arguments{
\item{df}{data.frame}

\item{time_colname}{}

\item{harmonics}{}

\item{period}{}

\item{sinusoid}{}

\item{n_knots}{}
}
\value{

}
\description{
Limorhyde function with multiple harmonics
}
