% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_analysis.R
\name{make_size_filter}
\alias{make_size_filter}
\title{Helper function to subset gene ontology lists}
\usage{
make_size_filter(low, high)
}
\arguments{
\item{low}{minimum number of genes}

\item{high}{maximum number of genes}
}
\value{
function to filter lists of genes
}
\description{
Helper function to subset gene ontology lists
}
