% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq_updated.R
\name{make_venn}
\alias{make_venn}
\title{Draws a Venn diagram of significant entities (genes or GO-terms) between tissues across contrasts,
or between contrasts across tissues}
\usage{
make_venn(between_contrasts, type, tissue_str, cond_str, genes_type, data_dir)
}
\arguments{
\item{between_contrasts}{boolean: Whether to plot overlaps between contrasts (TRUE) or between tissues (FALSE)}

\item{type}{character vector: Either "DE" or "GO" depending on the type of entity}

\item{tissue_str}{character vector: Comma-seperated list of tissues as present in the project root, i.e. Hyp, NAc}

\item{cond_str}{character vector: Comma-separated list of contrasts as present in the data directory, i.e. Group2,Group3}
}
\value{
A Venn Diagram of the ggplot-kind.
}
\description{
Draws a Venn diagram of significant entities (genes or GO-terms) between tissues across contrasts,
or between contrasts across tissues
}
