% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{merge_cov}
\alias{merge_cov}
\title{Merge list of cov data}
\usage{
merge_cov(cov_data)
}
\arguments{
\item{cov_data}{list of data.tables returned by prepare_cov_files. Elements
in cov_data are updated by reference when running this function.
Elements must be named. Names are used to construct column names in the
output.}
}
\value{
data.table with the rows present in all samples
}
\description{
Merge list of cov data
}
