% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_analysis.R
\name{merge_featurecounts}
\alias{merge_featurecounts}
\title{Read featureCounts output}
\usage{
merge_featurecounts(files, regex)
}
\arguments{
\item{regex}{optional regular expression for updating sample names.
Default option works with our pipeline. Set to NULL to disable.}

\item{x}{path to a file}
}
\value{
data.table with gene counts and reformatted sample names
}
\description{
Read featureCounts output
}
