% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisulfite_analysis.R
\name{meth_filter}
\alias{meth_filter}
\alias{meth_filter.DGEList}
\alias{meth_filter.data.frame}
\alias{meth_filter.data.table}
\title{Count filtering for methylation counts}
\usage{
meth_filter(y, min_counts = 8, filter_extremes = TRUE)

\method{meth_filter}{DGEList}(y, min_counts = 8, filter_extremes = TRUE)

\method{meth_filter}{data.frame}(y, min_counts = 8, filter_extremes = TRUE)

\method{meth_filter}{data.table}(y, min_counts = 8, filter_extremes = TRUE)
}
\arguments{
\item{y}{DGEList}

\item{min_counts}{minimum number of counts}

\item{filter_extremes}{logical, filter 0\% or 100\% in all}
}
\value{
filtered DGEList
}
\description{
Count filtering for methylation counts
}
\section{Methods (by class)}{
\itemize{
\item \code{DGEList}: Count filtering for methylation counts DGEList

\item \code{data.frame}: Count filtering for methylation counts data.frame

\item \code{data.table}: Count filtering for methylation counts data.table
}}

