% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{minus_log_trans}
\alias{minus_log_trans}
\title{Title}
\usage{
minus_log_trans(base = 10)
}
\arguments{
\item{base}{a positive or complex number: the base with respect to which
logarithms are computed. Defaults to 10.}
}
\value{
transformer, see scales::trans_new
}
\description{
Title
}
\examples{
minus_log_trans()
}
