% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian_analysis.R
\name{nearest_timepoint}
\alias{nearest_timepoint}
\title{Find nearest timepoint}
\usage{
nearest_timepoint(x, timepoints, cycle_length = 24)
}
\arguments{
\item{x}{vector, of new timepoints}

\item{timepoints}{vector, of target timepoints}

\item{cycle_length}{integer, length of a cycle. Typically 24 or 60.}
}
\value{
vector of same length as x, where the entries are the nearest value
supplied in 'timepoints'. In case of ties the first entry in 'timepoints'
is selected.
}
\description{
Find nearest timepoint
}
\examples{
nearest_timepoint(0:36, timepoints = 0:5 * 4)
}
