% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_figures.R
\name{plot_all_md}
\alias{plot_all_md}
\title{Prepare MD plots}
\usage{
plot_all_md(y)
}
\arguments{
\item{y}{DGEList}
}
\value{
One or more plots are created on the current device
}
\description{
Prepare MD plots
}
\examples{
library(edgeR)
set.seed(1)
counts <- matrix(sample(1:20), ncol = 4)
y <- DGEList(counts)
plot_all_md(y)
}
