% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_figures.R
\name{plot_sample_heatmap}
\alias{plot_sample_heatmap}
\title{Plot heatmap of sample distances}
\usage{
plot_sample_heatmap(x, method, cpm = FALSE, log = FALSE, ...)
}
\arguments{
\item{x}{a DGE list or a matrix of numbers}

\item{method}{string, method used for calculating sample distances. Supported methods
are poisson, MDS, and the methods supported by dist. See details.}

\item{cpm}{logical, should input be converted to counts per million, see details}

\item{log}{logical, should CPM be log-transformed, see details}

\item{...}{passed to pheatmap}
}
\value{

}
\description{
Plot heatmap of sample distances
}
\details{
If the method is MDS, the limma/edgeR plotMDS is used to calculate
sample distances using the first two dimensions. If the method is poisson distance is calculated using the
poiClaClu package. Otherwise input is optionally converted to (log)CPM and
passed to dist with the specified method.
}
