% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq.R
\name{plot_top_50_DE_genes_heatmap}
\alias{plot_top_50_DE_genes_heatmap}
\title{Plot top 50 DE genes}
\usage{
plot_top_50_DE_genes_heatmap(y, gene_test_list, order_by_vec)
}
\arguments{
\item{y}{DGE list object}

\item{gene_test_list}{Output from EdgeR tester}

\item{Character}{vector containing the variable to order by as present in the y$samples dataframe.}
}
\value{
Plots a heatmap
}
\description{
Plot top 50 DE genes
}
\examples{
plot_top_50_DE_genes_heatmap
}
