% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{prepare_featurecounts}
\alias{prepare_featurecounts}
\title{Prepare featureCounts output files for RNA-seq analysis}
\usage{
prepare_featurecounts(files, regex)
}
\arguments{
\item{files}{list of featureCounts output files}

\item{regex}{optional regular expression for updating sample names.
Default option works with the fq2count pipeline. Set to NULL to disable.}
}
\value{
data.table of gene counts and formatted gene information. FeatureCounts
reports chromosome, start, end and strand for each exon. These are reformatted
so start contains the start of the first exon and end the end of the last exon.
Only one copy of chromosome and strand are reported. Column names are also
reformatted to match with Bioconductor nomenclature (seqnames instead of Chr, etc.).
If a regex is supplied it is used to remove text from the sample names, if not
specified a regex that works with the fq2count pipeline is used.
}
\description{
Prepare featureCounts output files for RNA-seq analysis
}
\examples{
\dontrun{
featurecounts_files <- c(
  system.file("extdata", "KO_1.txt.gz", package = "cbmr"),
  system.file("extdata", "WT_1.txt.gz", package = "cbmr"),
  system.file("extdata", "KO_2.txt.gz", package = "cbmr"),
  system.file("extdata", "WT_2.txt.gz", package = "cbmr"),
  system.file("extdata", "KO_3.txt.gz", package = "cbmr"),
  system.file("extdata", "WT_3.txt.gz", package = "cbmr")
)
rna_counts <- prepare_featurecounts(featurecounts_files)
}
}
