% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_figures.R
\name{prepare_heatmap_data.default}
\alias{prepare_heatmap_data.default}
\title{Title}
\usage{
\method{prepare_heatmap_data}{default}(x, method, cpm, log)
}
\arguments{
\item{x}{}

\item{method}{}

\item{cpm}{}

\item{log}{}
}
\value{

}
\description{
Title
}
