% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_figures.R
\name{prepare_heatmap_data}
\alias{prepare_heatmap_data}
\alias{prepare_heatmap_data.default}
\alias{prepare_heatmap_data.DGEList}
\alias{prepare_heatmap_data.EList}
\title{Prepare data for heatmap}
\usage{
prepare_heatmap_data(x, method, cpm = FALSE, log = FALSE)

\method{prepare_heatmap_data}{default}(x, method, cpm = FALSE, log = FALSE)

\method{prepare_heatmap_data}{DGEList}(x, method, cpm = FALSE, log = FALSE)

\method{prepare_heatmap_data}{EList}(x, method, cpm = FALSE, log = FALSE)
}
\arguments{
\item{x}{object containing read counts, either a matrix, a DGEList object or an EList object}

\item{method}{character, method to calculate distances, one of 'MDS',
'poission' and the methods supported by stats::dist, see details.}

\item{cpm}{logical, converted to counts per million?
Only used for the methods supported by stats::dist}

\item{log}{logical, log transform?
Only used for the methods supported by stats::dist}
}
\value{
a 'dist' object with sample to sample distances
}
\description{
Prepare data for heatmap
}
\details{
Uses a number of methods to prepare data for creating a heatmap. x should
contain read counts. If the method 'MDS' or 'poisson' is selected the raw counts
are processed with limma::plotMDS or PoiClaClu::PoissonDistance, otherwise counts
are optionally converted to counts per million and/or log transformed before
distances are calculated with stats::dist.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: prepare heatmap data default

\item \code{DGEList}: prepare heatmap data for DGEList

\item \code{EList}: prepare heatmap data for DGEList
}}

\examples{
library(edgeR)
set.seed(1)
counts <- matrix(sample(1:20), ncol = 4)
prepare_heatmap_data(counts, method = "euclidean")
}
