% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_figures.R
\name{prepare_mds_data.default}
\alias{prepare_mds_data.default}
\title{Prepare data for MDS plot}
\usage{
\method{prepare_mds_data}{default}(y, dim_plot, colour_by = NULL)
}
\arguments{
\item{y}{}

\item{dim_plot}{}

\item{colour_by}{}

\item{col_scale}{}
}
\value{

}
\description{
Prepare data for MDS plot
}
