% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_figures.R
\name{prepare_mds_data}
\alias{prepare_mds_data}
\title{Prepare data for MDS plot}
\usage{
prepare_mds_data(y, ...)
}
\arguments{
\item{y}{}

\item{...}{}
}
\value{

}
\description{
Prepare data for MDS plot
}
