% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{prepare_metadata}
\alias{prepare_metadata}
\title{Prepare metadata data.tale from excel file}
\usage{
prepare_metadata(
  file,
  new_col_names = NULL,
  uninteresting_cols = NULL,
  use_defult_new_names = TRUE,
  use_default_uninteresting = TRUE
)
}
\arguments{
\item{file}{string, path to excel file}

\item{new_col_names}{optional named character vector, see details}

\item{uninteresting_cols}{optional character vector, see details}

\item{use_defult_new_names}{boolean, use the default name conversions found
in the variable computer_friendly_names?}

\item{use_default_uninteresting}{boolean, use the default list of
uninteresting columns found in the variable uninteresting_names?}
}
\value{
data.table with formatted metadata
}
\description{
Prepare metadata data.tale from excel file
}
\details{
new_col_names is a named vector of the format
c("excel col name" = "updated_name"). An example is included in the variable
computer_friendly_names.

uninteresting_cols is a vector of columns that will be removed. This removal
happens after column names are updated, so if a column is updated with
new_col_names, the updated name needs to be given in uninteresting_cols.
A default list is included in the variable uninteresting_names.
}
