% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{prepare_rna_dgelist}
\alias{prepare_rna_dgelist}
\title{Prepare a DGEList object}
\usage{
prepare_rna_dgelist(counts, metadata, sample_col = "Sample ID")
}
\arguments{
\item{counts}{data.table with gene counts, if data.table has keys those are used to determine gene information}

\item{metadata}{data.table with metadata on the samples}

\item{sample_col}{column in metadata that corresponds to sample (column) names in counts}
}
\value{
DGEList object with information on both genes and samples
}
\description{
Prepare a DGEList object
}
\examples{
\dontrun{
  prepare_rna_dgelist(counts = rna_counts, 
                      metadata = example_metadata, 
                      sample_col = "Sample.ID"
                      )
}
}
