% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{prepare_rrbs_dgelist}
\alias{prepare_rrbs_dgelist}
\title{Prepare a DGEList object for RRBS analysis}
\usage{
prepare_rrbs_dgelist(
  counts,
  metadata,
  sample_col = "Sample.ID",
  aggregate_by = NULL
)
}
\arguments{
\item{counts}{data.table with gene counts}

\item{metadata}{data.table with metadata on the samples}

\item{sample_col}{column in metadata that corresponds to sample (column) names in counts}
}
\value{
DGEList object with information on both genes and samples
}
\description{
Prepare a DGEList object for RRBS analysis
}
