% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{preprocess_methylation_data}
\alias{preprocess_methylation_data}
\title{Wrapper function for preprocessing methylation data}
\usage{
preprocess_methylation_data(
  x,
  remove = NULL,
  select = NULL,
  genes = NULL,
  annotations = NULL
)
}
\arguments{
\item{x}{data.table with methylation information.}

\item{remove}{optional, list of files with regions to remove}

\item{select}{optional, list of files with regions to select}

\item{genes}{optional, gtf file containing gene information}

\item{annotations}{optional, list of files containing additional annotation information.
If the list is named, only the column with that name will be used. Mixing named
and unnamed elements is supported.}
}
\value{

}
\description{
Wrapper function for preprocessing methylation data
}
\details{
The files in remove, select and annotation must be
(optionally gzipped) csv files, or one of the filetypes supported by \link[rtracklayer]{import}.
If file is a csv file, the first three columns must be seqnames, start and end.
}
