% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq.R
\name{print_DE_results}
\alias{print_DE_results}
\title{Prints the HTML code to write in the HTML in an asis-chunk, containing the tables of resultss}
\usage{
print_DE_results(test_list, contrast, type = "edgeR")
}
\arguments{
\item{type}{The type of dataframe}

\item{df}{The DE dataframe}
}
\value{
Raw HTML code
}
\description{
Prints the HTML code to write in the HTML in an asis-chunk, containing the tables of resultss
}
\examples{
print_DE_results
}
