% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_bulk_RNA_seq.R
\name{print_DT}
\alias{print_DT}
\title{Auxillary function that prints a DT::datatable object correctly
both when working interactively and in reports}
\usage{
print_DT(DT)
}
\arguments{
\item{DT}{The DT dataframe to plot}
}
\value{
Nothing, invisible object
}
\description{
Auxillary function that prints a DT::datatable object correctly
both when working interactively and in reports
}
