% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{remove_shallow_samples}
\alias{remove_shallow_samples}
\title{Remove samples with few sites covered}
\usage{
remove_shallow_samples(cov_data, min_sites)
}
\arguments{
\item{cov_data}{list of data.tables returned by prepare_cov_files}

\item{min_sites}{minimum number of sites covered}
}
\value{
list of data.tables with min_sites or more covered CpGs
}
\description{
Remove samples with few sites covered
}
