% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq.R
\name{render_markdown_file_single_tissue}
\alias{render_markdown_file_single_tissue}
\title{Render one markdown report for a single tissue}
\usage{
render_markdown_file_single_tissue(markdown_path, tissue, exclude)
}
\arguments{
\item{markdown_path}{The path of the markdown file}

\item{tissue}{The vector to filter the file by}

\item{exclude}{The samples to exclude when testing}
}
\value{
Nothing
}
\description{
Render one markdown report for a single tissue
}
\examples{
render_markdown_file_single_tissue()
}
