% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq_updated.R
\name{render_markdown_file_single_tissue_updated}
\alias{render_markdown_file_single_tissue_updated}
\title{Render a single tissue}
\usage{
render_markdown_file_single_tissue_updated(
  markdown_path,
  tissue,
  exclude,
  suffix = ""
)
}
\arguments{
\item{markdown_path}{The absolute path to the markdown fle}

\item{tissue}{Character vector of length 1, indicating the tissue to keep,
as present in the metadata sheet for the project}

\item{exclude}{Character vector of length 1, indicating the Sample IDS to
remove seperated by commas, as present in the metadata sheet for the project}
}
\value{
Returns nothing, but creates the corresponding .html files with descriptive names
}
\description{
Render a single tissue
}
