% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq.R
\name{render_tissues}
\alias{render_tissues}
\title{Render markdown reports for multiple tissues}
\usage{
render_tissues(markdown_path, tissue_exclusion_vec)
}
\arguments{
\item{markdown_path}{The path of the template markdown file}

\item{tissue}{The vector to filter the file by}

\item{exclude}{The samples to exclude when testing}
}
\value{
Nothing
}
\description{
Render markdown reports for multiple tissues
}
\examples{
render_markdown_file_single_tissue()
}
