% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_bulk_RNA_seq.R
\name{replace_na_with_INT}
\alias{replace_na_with_INT}
\title{Inverse-normal transforms a vector with NA's in it}
\usage{
replace_na_with_INT(vec)
}
\arguments{
\item{vec}{The vector to normal transform}
}
\value{
The inverse-normal transformed vector
}
\description{
Inverse-normal transforms a vector with NA's in it
}
