% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rna_counts}
\alias{rna_counts}
\title{Small example of a data frame with RNA counts}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 200 rows and 12 columns.
}
\usage{
rna_counts
}
\description{
A dataset with a small toy example of what RNA counts look like af reading in
}
\keyword{datasets}
