% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_analysis.R
\name{run_ontology_tests}
\alias{run_ontology_tests}
\title{Run ontology tests and annotate results}
\usage{
run_ontology_tests(terms, y, contrast_matrix = NULL, coefs = NULL, fun, ...)
}
\arguments{
\item{terms}{list with two elements: 'index' containing an index vector as
described in limma::camera and 'annotation' containing information on the
ontologies tested. Can be generated with get_enrichment_terms}

\item{y}{DGEList object to be tested}

\item{contrast_matrix}{matrix with contrasts to be tested. Both contrast_matrix
and coefs can be specified simultaneously.}

\item{coefs}{coefficeints of y$design to be tested Both contrast_matrix
and coefs can be specified simultaneously.}

\item{fun}{function to use for testing. Typically limma::camera or limma::fry}

\item{...}{other parameters passed to fun.}
}
\value{
list of enrichment results
}
\description{
Run ontology tests and annotate results
}
\examples{
\dontrun{
library(org.Mm.eg.db)
y <- prepare_rna_dgelist(rna_counts, example_metadata, sample_col = "Sample.ID")
design <- model.matrix(~ 0 + Genotype + Batch, data = y$samples)
y <- edgeR::estimateDisp(y, design = design)
rownames(y) <- y$genes$Geneid
contrast_matrix <- limma::makeContrasts(Genotype = GenotypeKO - GenotypeWT, 
                                        levels = y$design)
terms <- get_enrichment_terms(org.Mm.eg.db, y$genes$Geneid)
run_ontology_tests(terms$BP, y, contrast_matrix, c("Batch2", "Batch3"), limma::camera)
}
}
