% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_RNA_seq_updated.R
\name{scale_this}
\alias{scale_this}
\title{Scaling function that does not return weird matrix}
\usage{
scale_this(vec)
}
\arguments{
\item{vec}{Vector to scale}
}
\value{
Scaled vector
}
\description{
Scaling function that does not return weird matrix
}
