% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{symlog_trans}
\alias{symlog_trans}
\title{Symlog transformation}
\usage{
symlog_trans(base = 10, thr = 1, scale = 1)
}
\arguments{
\item{base}{Exponential base, default is 10}

\item{thr}{Threshold for linearity, default is 1}

\item{scale}{Scale factor applied to the log transformed data}
}
\description{
A log scale that handles negative values and is linear around 0. Similar to
the symlog from Python.
Copied from: https://stackoverflow.com/questions/14613355/how-to-get-something-like-matplotlibs-symlog-scale-in-ggplot-or-lattice
}
