% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{uninteresting_names}
\alias{uninteresting_names}
\title{Metadata uninteresting names}
\format{
An object of class \code{character} of length 17.
}
\usage{
uninteresting_names
}
\description{
List of metadata column names that are rarely needed for differential
expression analysis
}
\keyword{datasets}
