% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{covfile_reader}
\alias{covfile_reader}
\title{Read cov file and optionally aggregate to known CpGs}
\usage{
covfile_reader(file, cpgs = NULL)
}
\arguments{
\item{file}{}

\item{cpgs}{}
}
\value{

}
\description{
Read cov file and optionally aggregate to known CpGs
}
