% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{featurecounts_reader}
\alias{featurecounts_reader}
\title{Read featureCounts output}
\usage{
featurecounts_reader(x, regex = NULL)
}
\arguments{
\item{x}{path to a file}

\item{regex}{optional regular expression for updating sample names.
Default option works with the fq2count pipeline. Set to NULL to disable.}
}
\value{
data.table with gene counts and reformatted sample names
}
\description{
Read featureCounts output
}
