% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisulfite_specific.R
\name{get_delta_methylation}
\alias{get_delta_methylation}
\title{Get delta methylation metric}
\usage{
get_delta_methylation(group_methylation, contrasts)
}
\arguments{
\item{group_methylation}{object coercible to a matrix containing groupwise
methylation (or effect sizes) as calcualted by get_groupwise_methylation}

\item{contrasts}{matrix of contrasts as created by edgeR::makeContrasts}
}
\value{
matrix with the effect sizes of contrasts
}
\description{
Calculate effect sizes of contrasts
}
\examples{
set.seed(1)
n_samples <- 4
n_genes <- 10
counts <- matrix(sample(1:20, n_samples \* 2 \* n_genes, replace = TRUE), 
                 ncol = n_samples * 2)
colnames(counts) <- paste0("sample", rep(1:4, each = 2), c("_Me", "_Un"))
methylation_levels <- get_methylation_metric(counts)

group <- c("a", "a", "b", "b")
design <- model.matrix(~ 0 + grp)

group_methylation <- get_groupwise_methylation(methylation_levels, design)

contrasts <- edgeR::makeContrasts(grpb - grpa, (grpb + grpa)/2, levels = design)
get_delta_methylation(group_methylation, contrasts)
}
