% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_figures.R
\name{ggplot_mds}
\alias{ggplot_mds}
\title{Plot MDS plot}
\usage{
ggplot_mds(y, dim_plot, colour_by = NULL, col_scale)
}
\arguments{
\item{y}{}

\item{dim_plot}{}

\item{colour_by}{}

\item{col_scale}{}
}
\value{

}
\description{
Plot MDS plot
}
