% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_figures.R
\name{plot_density}
\alias{plot_density}
\title{Prepare density plots}
\usage{
plot_density(y)
}
\arguments{
\item{y}{DGEList to be plotted}
}
\value{
One or more plots are created on the current device,
list of the plot generated are also returned
}
\description{
Prepare density plots
}
\examples{
\dontrun{
  p <- plot_density(y)
  p
}
}
