% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_figures.R
\name{plot_retained_cpgs}
\alias{plot_retained_cpgs}
\title{Plot number of CpGs retained by including samples.}
\usage{
plot_retained_cpgs(x)
}
\arguments{
\item{x}{list of data.tables with CpG information}
}
\value{
ggplot object that creates the plot
}
\description{
Plot number of CpGs retained by including samples.
}
\examples{
plot_retained_cpgs(cov_list)
}
