% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_bulk_RNA_seq.R
\name{print_plotly}
\alias{print_plotly}
\title{Auxillary function that plots a GGplotly object correctly
both when working interactively and in reports}
\usage{
print_plotly(plotly_plot)
}
\arguments{
\item{plotly_plot}{The ggplotly-object to plot}
}
\value{
Nothing, invisible object
}
\description{
Auxillary function that plots a GGplotly object correctly
both when working interactively and in reports
}
